+++ src/nifigator/__init__.py
@@ -27,3 +27,4 @@ from .pdfparser import *
 from .utils import *
 from .lemongraph import *
 from .lemonobjects import *
+from .set_math import *