+++ src/nifigator/nifvecobjects.py
@@ -13,29 +13,38 @@ from rdflib.term import IdentifiedNode, URIRef, Literal
 from rdflib.plugins.stores import sparqlstore, memory
 from rdflib.namespace import NamespaceManager
 from .const import (
+    STOPWORDS,
     RDF,
     XSD,
     NIF,
     NIFVEC,
+    ONTOLEX,
+    LEXINFO,
+    DECOMP,
     DEFAULT_URI,
     DEFAULT_PREFIX,
     MIN_PHRASE_COUNT,
     MIN_CONTEXT_COUNT,
     MIN_PHRASECONTEXT_COUNT,
     MAX_PHRASE_LENGTH,
-    MAX_LEFT_LENGTH,
-    MAX_RIGHT_LENGTH,
-    MIN_LEFT_LENGTH,
-    MIN_RIGHT_LENGTH,
+    MAX_CONTEXT_LENGTH,
+    TRIPLE_BATCH_SIZE,
     CONTEXT_SEPARATOR,
     PHRASE_SEPARATOR,
     WORDS_FILTER,
     FORCED_SENTENCE_SPLIT_CHARACTERS,
-    DEFAULT_CONTEXT_SEPARATOR,
-    DEFAULT_PHRASE_SEPARATOR,
 )
 from .utils import tokenizer, tokenize_text, to_iri
 from .nifgraph import NifGraph
+from .set_math import merge_multiset
+
+default_min_phrase_count = 2
+default_min_phrasecontext_count = 2
+default_min_context_count = 2
+default_max_context_length = 5
+default_max_phrase_length = 5
+default_context_separator = "_"
+default_phrase_separator = "+"
 
 
 class NifVectorGraph(NifGraph):
@@ -44,15 +53,13 @@ class NifVectorGraph(NifGraph):
 
     :param nif_graph: the graph from which to construct the NIF Vector graph (optional)
 
-    :param context_uris: the context uris of the contexts used with the nif_graph to construct the NIF Vector graph
+    :param context_uris: the context uris of the contexts used with the nif_graph to construct the NIF Vector graph (optional)
 
     :param documents: the documents from which to construct the NIF Vector graph (optional)
 
-    :param lexicon: the namespace of the lexicon (words and phrases in the text)
+    :param base_uri: the namespace of the nifvec data
 
-    :param window: the namespace of the windows (context-phrase combinations in the text)
-
-    :param context: the namespace of the contexts
+    :param lang: the language of the nifvec data
 
     :param params: dictionary with parameters for constructing the NIF Vector graph
 
@@ -62,10 +69,9 @@ class NifVectorGraph(NifGraph):
         self,
         nif_graph: NifGraph = None,
         context_uris: list = None,
-        sentences: list = None,
-        lexicon: Namespace = Namespace(DEFAULT_URI + "lexicon/"),
-        window: Namespace = Namespace(DEFAULT_URI + "window/"),
-        context: Namespace = Namespace(DEFAULT_URI + "context/"),
+        documents: list = None,
+        base_uri: Namespace = Namespace(DEFAULT_URI + "nifvec-data/"),
+        lang: str = None,
         params: dict = {},
         store: Union[Store, str] = "default",
         identifier: Optional[Union[IdentifiedNode, str]] = None,
@@ -91,162 +97,140 @@ class NifVectorGraph(NifGraph):
         else:
             self.params[WORDS_FILTER] = None
 
-        self.bind("lexicon", lexicon)
-        self.bind("window", window)
-        self.bind("context", context)
+        self.base_uri = base_uri
+        self.lang = lang
+        self.bind("nifvec-data", base_uri)
         self.bind("nifvec", NIFVEC)
         self.bind("nif", NIF)
+        self.bind("ontolex", ONTOLEX)
+        self.bind("lexinfo", LEXINFO)
+        self.bind("decomp", DECOMP)
 
         if nif_graph is not None:
-            logging.info(".. Extracting text from graph")
-            sentences = dict()
+            # if nif_graph is available then contexts are extracted from this graph
+            logging.debug(".. extracting documents from graph")
+            documents = dict()
             contexts = nif_graph.contexts
             for context in contexts:
+                # if context_uris is None then all contexts are extracted
+                # otherwise only those in the context_uris list
                 if context_uris is None or context.uri in context_uris:
-                    if context.sentences is not None:
-                        for sent in context.sentences:
-                            sentences[sent.uri] = sent.anchorOf
+                    isString = context.isString
+                    if isString is not None:
+                        documents[context.uri] = preprocess(isString)
                     else:
-                        logging.warning("No sentences found for " + str(context.uri))
+                        logging.warning("No isString found for " + str(context.uri))
 
-        if sentences is not None:
-            logging.info(".. Creating windows dict")
-            windows = generate_windows(documents=sentences, params=self.params)
-            logging.info(".. Creating phrase and context dicts")
-            phrase_count, context_count = self.create_window_phrase_count_dicts(
-                windows=windows
+        if documents is not None:
+            phrases = generate_document_phrases(documents=documents, params=self.params)
+            contexts, phrases = generate_document_contexts(
+                init_phrases=phrases, documents=documents, params=self.params
             )
-            logging.info(".. Collecting triples")
-            for triple in self.generate_triples(
-                windows=windows,
-                phrase_count=phrase_count,
-                context_count=context_count,
-            ):
-                self.add(triple)
-            logging.info(".. Finished initialization")
+            self.store_triples(
+                phrases=phrases,
+                contexts=contexts,
+            )
+
+    def store_triples(
+        self,
+        phrases: dict = {},
+        contexts: dict = {},
+    ):
+        """
+        Function to store the triples from a document set into the NifVector graph.
+
+        The triples are loaded in batches into the NifVector graph, to prevent the number of SPARQL updates.
+
+        :param phrases: dictionary of all phrases to be stored
+
+        :param contexts: dictionary of all contexts to be stored
+
+        """
+        triple_batch_size = self.params.get(TRIPLE_BATCH_SIZE, 5e6)
+        count = 1
+        temp_g = Graph()
+        for triple in self.generate_triples(phrases=phrases, contexts=contexts):
+            temp_g.add(triple)
+            if count == triple_batch_size:
+                self += temp_g
+                count = 1
+                temp_g = Graph()
+            else:
+                count += 1
+        self += temp_g
+        logging.debug(".. finished storing triples")
 
     def generate_triples(
         self,
-        windows: dict = {},
-        phrase_count: dict = {},
-        context_count: dict = {},
+        phrases: dict = {},
+        contexts: dict = {},
     ):
-        """ """
-        context_sep = self.params.get(CONTEXT_SEPARATOR, DEFAULT_CONTEXT_SEPARATOR)
-        phrase_sep = self.params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
-        ns = dict(self.namespaces())
-        lexicon_ns = ns["lexicon"]
-        context_ns = ns["context"]
-        window_ns = ns["window"]
+        """
+        Function to create all triples of a set of documents
 
-        for phrase in phrase_count.keys():
-            phrase_text = to_iri(phrase)
-            phrase_value = Literal(phrase_text, datatype=XSD.string)
-            phrase_uri = URIRef(lexicon_ns + phrase_text)
-            count = Literal(phrase_count[phrase], datatype=XSD.nonNegativeInteger)
+        :param phrases: dictionary of all phrases to be stored
+
+        :param contexts: dictionary of all contexts to be stored
+
+        """
+
+        logging.debug(".. collecting triples")
+
+        context_sep = self.params.get(CONTEXT_SEPARATOR, default_context_separator)
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
+
+        # to add: nifvec graph definition, which contexts? which language, stopwords
+
+        for phrase, value in phrases.items():
+            phrase_uri = URIRef(self.base_uri + to_iri(phrase))
+            phrase_value = Literal(phrase.replace(phrase_sep, " "), datatype=XSD.string)
+            count = Literal(value, datatype=XSD.nonNegativeInteger)
             yield ((phrase_uri, RDF.type, NIF.Phrase))
             yield ((phrase_uri, RDF.value, phrase_value))
             yield ((phrase_uri, NIFVEC.hasCount, count))
-        for context in context_count.keys():
-            context_text = to_iri(context[0]) + context_sep + to_iri(context[1])
-            context_value = Literal(context_text, datatype=XSD.string)
-            left_context_value = Literal(to_iri(context[0]), datatype=XSD.string)
-            right_context_value = Literal(to_iri(context[1]), datatype=XSD.string)
-            context_uri = URIRef(context_ns + context_text)
-            count = Literal(context_count[context], datatype=XSD.nonNegativeInteger)
+        logging.debug(".... finished triples for phrases")
+
+        for ((left_part, right_part)), value in contexts.items():
+            context_uri = URIRef(
+                self.base_uri + to_iri(left_part) + context_sep + to_iri(right_part)
+            )
+            left_context_value = Literal(
+                left_part.replace(phrase_sep, " "), datatype=XSD.string
+            )
+            right_context_value = Literal(
+                right_part.replace(phrase_sep, " "), datatype=XSD.string
+            )
+            context_count = Literal(
+                sum(v for v in value.values()), datatype=XSD.nonNegativeInteger
+            )
             yield ((context_uri, RDF.type, NIFVEC.Context))
             yield ((context_uri, NIFVEC.hasLeftValue, left_context_value))
             yield ((context_uri, NIFVEC.hasRightValue, right_context_value))
-            yield ((context_uri, RDF.value, context_value))
-            yield ((context_uri, NIFVEC.hasCount, count))
-        for phrase in windows.keys():
-            for window in windows[phrase]:
-                p = to_iri(window[0])
-                n = to_iri(window[1])
-                phrase_text = to_iri(phrase)
-                phrase_uri = URIRef(lexicon_ns + phrase_text)
-                context_uri = URIRef(context_ns + p + context_sep + n)
-                window_text = p + context_sep + phrase_text + context_sep + n
-                window_value = Literal(window_text, datatype=XSD.string)
-                window_uri = URIRef(window_ns + window_text)
-                window_count = Literal(
-                    windows[phrase][window], datatype=XSD.nonNegativeInteger
+            yield ((context_uri, NIFVEC.hasCount, context_count))
+        logging.debug(".... finished triples for contexts")
+
+        for ((left_part, right_part)), value in contexts.items():
+            context_uri = URIRef(
+                self.base_uri + to_iri(left_part) + context_sep + to_iri(right_part)
+            )
+            for phrase, phrase_value in value.items():
+                window_uri = URIRef(
+                    self.base_uri
+                    + to_iri(left_part)
+                    + context_sep
+                    + to_iri(phrase)
+                    + context_sep
+                    + to_iri(right_part)
                 )
-                yield ((phrase_uri, NIFVEC.isPhraseOf, window_uri))
-                yield ((context_uri, NIFVEC.isContextOf, window_uri))
+                phrase_uri = URIRef(self.base_uri + to_iri(phrase))
+                window_count = Literal(phrase_value, datatype=XSD.nonNegativeInteger)
                 yield ((window_uri, RDF.type, NIFVEC.Window))
-                yield ((window_uri, RDF.value, window_value))
                 yield ((window_uri, NIFVEC.hasContext, context_uri))
                 yield ((window_uri, NIFVEC.hasPhrase, phrase_uri))
                 yield ((window_uri, NIFVEC.hasCount, window_count))
-
-    def create_window_phrase_count_dicts(self, windows: dict = {}):
-        """
-        Function to prune and create phrase dict and contexts dict
-        """
-        min_phrasecontext_count = self.params.get("min_phrasecontext_count", 5)
-        min_phrase_count = self.params.get("min_phrase_count", 5)
-        min_context_count = self.params.get("min_context_count", 5)
-
-        # delete phrasewindow with number of occurrence < MIN_PHRASECONTEXT_COUNT
-        to_delete = []
-        for d_phrase in windows.keys():
-            for d_window in windows[d_phrase].keys():
-                if windows[d_phrase][d_window] < min_phrasecontext_count:
-                    to_delete.append((d_phrase, d_window))
-        logging.info(
-            ".... deleting "
-            + str(len(to_delete))
-            + " windows from "
-            + str(
-                len(
-                    [
-                        d_window
-                        for d_phrase in windows.keys()
-                        for d_window in windows[d_phrase].keys()
-                    ]
-                )
-            )
-        )
-        for item in to_delete:
-            del windows[item[0]][item[1]]
-
-        # create context_count and phrase_count
-        context_count = defaultdict(int)
-        phrase_count = defaultdict(int)
-        for d_phrase in windows.keys():
-            for d_window in windows[d_phrase].keys():
-                c = windows[d_phrase][d_window]
-                context_count[d_window] += c
-                phrase_count[d_phrase] += c
-
-        # delete phrases with number of occurrence < MIN_PHRASE_COUNT
-        to_delete = [
-            p for p in phrase_count.keys() if phrase_count[p] < min_phrase_count
-        ]
-        logging.info(
-            ".... deleting "
-            + str(len(to_delete))
-            + " phrases from "
-            + str(len(phrase_count.keys()))
-        )
-        for item in to_delete:
-            del phrase_count[item]
-
-        # delete windows with number of occurrence < MIN_CONTEXT_COUNT
-        to_delete = [
-            c for c in context_count.keys() if context_count[c] < min_context_count
-        ]
-        logging.info(
-            ".... deleting "
-            + str(len(to_delete))
-            + " contexts from "
-            + str(len(context_count.keys()))
-        )
-        for item in to_delete:
-            del context_count[item]
-
-        return phrase_count, context_count
+                yield ((phrase_uri, NIFVEC.isPhraseOf, window_uri))
+                yield ((context_uri, NIFVEC.isContextOf, window_uri))
+        logging.debug(".... finished triples for windows")
 
     def phrase_contexts(
         self,
@@ -270,17 +254,14 @@ class NifVectorGraph(NifGraph):
         :param topn: restrict output to topn (default = 15)
 
         """
-        ns = dict(self.namespaces())
-        lexicon_ns = ns["lexicon"]
-
-        context_sep = self.params.get(CONTEXT_SEPARATOR, DEFAULT_CONTEXT_SEPARATOR)
-        phrase_sep = self.params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
+        context_sep = self.params.get(CONTEXT_SEPARATOR, default_context_separator)
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
 
         if phrase_uri is None:
-            phrase_uri = URIRef(lexicon_ns + phrase_sep.join(phrase.split(" ")))
+            phrase_uri = URIRef(self.base_uri + phrase_sep.join(phrase.split(" ")))
 
         q = """
-    SELECT DISTINCT ?v (sum(?count) as ?n)
+    SELECT DISTINCT ?value_left ?value_right (sum(?count) as ?n)
     WHERE
     {\n"""
         q += (
@@ -292,24 +273,24 @@ class NifVectorGraph(NifGraph):
             ?w rdf:type nifvec:Window .
             ?w nifvec:hasContext ?c .
             ?w nifvec:hasCount ?count .
-            ?c rdf:value ?v .
             """
         )
         if left is not None:
             q += '?c nifvec:hasLeftValue "' + Literal(left) + '" .\n'
+        q += "?c nifvec:hasLeftValue ?value_left .\n"
         if right is not None:
             q += '?c nifvec:hasRightValue "' + Literal(right) + '" .\n'
-
+        q += "?c nifvec:hasRightValue ?value_right .\n"
         q += """
         }
     }
-    GROUP BY ?v
+    GROUP BY ?value_left ?value_right
     ORDER BY DESC(?n)
     """
         if topn is not None:
             q += "LIMIT " + str(topn) + "\n"
         results = Counter(
-            {tuple(r[0].split(context_sep)): r[1].value for r in self.query(q)}
+            {tuple([r[0].value, r[1].value]): r[2].value for r in self.query(q)}
         )
         return results
 
@@ -319,6 +300,8 @@ class NifVectorGraph(NifGraph):
         phrase_uri: URIRef = None,
         context: str = None,
         context_uri: URIRef = None,
+        contexts: list = None,
+        contexts_uris: list = None,
         topn: int = 15,
         topcontexts: int = 25,
         topphrases: int = 25,
@@ -332,39 +315,53 @@ class NifVectorGraph(NifGraph):
 
         :param context: the context to take into account for deriving similar phrases (as a string)
 
-        :param context: the context to take into account for deriving similar phrases (as a uri)
+        :param context_uri: the context to take into account for deriving similar phrases (as a uri)
+
+        :param contexts: use list of contexts to filter
+
+        :param contexts_uris: filter contexts
 
         :param topn: restrict output to topn (default = 15)
 
-        :param topcontexts: number of similar contexts to use
+        :param topcontexts: number of similar contexts to use when using phrase or phrase_uri
 
-        :param topphrases: number of similar phrases to use
+        :param topphrases: number of similar phrases to use when using context or context_uri
 
         """
-        ns = dict(self.namespaces())
-        lexicon_ns = ns["lexicon"]
-        phrase_sep = self.params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
-        if phrase_uri is None:
-            phrase_uri = URIRef(lexicon_ns + phrase_sep.join(phrase.split(" ")))
-        if context_uri is None and context is not None:
-            context_ns = ns["context"]
-            context_sep = self.params.get(CONTEXT_SEPARATOR, DEFAULT_CONTEXT_SEPARATOR)
-            context_uri = URIRef(context_ns + context_sep.join(context))
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
+        context_sep = self.params.get(CONTEXT_SEPARATOR, default_context_separator)
+        if phrase is not None:
+            phrase_uri = URIRef(self.base_uri + phrase_sep.join(phrase.split(" ")))
+        if context is not None:
+            context_uri = URIRef(
+                self.base_uri
+                + context_sep.join([c.replace(" ", phrase_sep) for c in context])
+            )
+        if contexts is not None:
+            contexts_uris = [
+                URIRef(
+                    self.base_uri
+                    + context_sep.join([c.replace(" ", phrase_sep) for c in context])
+                )
+                for context in contexts
+            ]
 
         q = """
     SELECT distinct ?v (count(?c) as ?num1)
     WHERE
     {\n"""
-        q += (
-            """
-        {
+        q += """
+        {"""
+        if phrase_uri is not None:
+            q += (
+                """
             {
                 SELECT DISTINCT ?c (sum(?count1) as ?n1) 
                 WHERE
                 {
                     """
-            + phrase_uri.n3()
-            + """ 
+                + phrase_uri.n3()
+                + """ 
                         nifvec:isPhraseOf ?w1 .
                     ?w1 rdf:type nifvec:Window .
                     ?w1 nifvec:hasContext ?c .
@@ -373,11 +370,11 @@ class NifVectorGraph(NifGraph):
                 GROUP BY ?c
                 ORDER BY DESC(?n1)
                 LIMIT """
-            + str(topcontexts)
-            + """
+                + str(topcontexts)
+                + """
             }
             """
-        )
+            )
         if context_uri is not None:
             q += (
                 """
@@ -405,7 +402,16 @@ class NifVectorGraph(NifGraph):
             ?p nifvec:isPhraseOf ?w .
             ?c nifvec:isContextOf ?w .
             ?w rdf:type nifvec:Window .
-            ?p rdf:value ?v .
+            ?p rdf:value ?v ."""
+
+        if contexts_uris is not None:
+            q += "FILTER (?c IN ("
+            for contexts_uri in contexts_uris:
+                q += contexts_uri.n3()
+                if contexts_uri != contexts_uris[-1]:
+                    q += ", "
+            q += "))"
+        q += """
         }
     }
     GROUP BY ?v
@@ -416,17 +422,15 @@ class NifVectorGraph(NifGraph):
         results = [item for item in self.query(q)]
         if len(results) > 0:
             norm = results[0][1].value
-            results = dict(
-                {r[0].replace(phrase_sep, " "): (r[1].value, norm) for r in results}
-            )
+            results = dict({r[0].value: (r[1].value, norm) for r in results})
         else:
             results = dict()
         return results
 
     def extract_rdf_type(self, rdf_type: str = None, topn: int = None):
         """ """
-        context_sep = self.params.get(CONTEXT_SEPARATOR, DEFAULT_CONTEXT_SEPARATOR)
-        phrase_sep = self.params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
+        context_sep = self.params.get(CONTEXT_SEPARATOR, default_context_separator)
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
         q = """
     SELECT distinct ?v (sum(?count) as ?num)
     WHERE
@@ -456,15 +460,26 @@ class NifVectorGraph(NifGraph):
 
     def phrases(self, topn: int = None):
         """
-        Returns phrases in the graph
+        Returns phrases with their counts in the graph
         """
-        return self.extract_rdf_type(rdf_type="nif:Phrase", topn=topn)
-
-    def windows(self, topn: int = None):
+        q = """
+    SELECT distinct ?v (sum(?count) as ?num)
+    WHERE
+    {\n"""
+        q += """
+        {
+            ?w rdf:type nif:Phrase .
+            ?w nifvec:hasCount ?count .
+            ?w rdf:value ?v .
+        }
+    }
+    GROUP BY ?v
+    ORDER BY DESC (?num)
         """
-        Returns windows in the graph
-        """
-        return self.extract_rdf_type(rdf_type="nifvec:Window", topn=topn)
+        if topn is not None:
+            q += "LIMIT " + str(topn) + "\n"
+        results = Counter({r[0].value: r[1].value for r in self.query(q)})
+        return results
 
     def dict_phrases_contexts(
         g, word: str = None, topn: int = 7, topcontexts: int = 10
@@ -489,16 +504,13 @@ class NifVectorGraph(NifGraph):
         Function that returns the phrases of a context
 
         """
-        ns = dict(self.namespaces())
-        context_ns = ns["context"]
-
-        context_sep = self.params.get(CONTEXT_SEPARATOR, DEFAULT_CONTEXT_SEPARATOR)
-        phrase_sep = self.params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
+        context_sep = self.params.get(CONTEXT_SEPARATOR, default_context_separator)
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
         context = (
             phrase_sep.join(context[0].split(" ")),
             phrase_sep.join(context[1].split(" ")),
         )
-        context_uri = URIRef(context_ns + context_sep.join(context)).n3()
+        context_uri = URIRef(self.base_uri + context_sep.join(context)).n3()
         q = """
     SELECT distinct ?v (sum(?s) as ?num)
     WHERE
@@ -521,9 +533,7 @@ class NifVectorGraph(NifGraph):
         )
         if topn is not None:
             q += "LIMIT " + str(topn) + "\n"
-        results = Counter(
-            {r[0].replace(phrase_sep, " "): r[1].value for r in self.query(q)}
-        )
+        results = Counter({r[0].value: r[1].value for r in self.query(q)})
         return results
 
     def compact(self):
@@ -564,110 +574,339 @@ class NifVectorGraph(NifGraph):
         logging.info(".. finished")
         return None
 
+    def find_otherForms(
+        self,
+        phrase: str = None,
+        phrase_uri: URIRef = None,
+    ):
+        """ """
+        phrase_sep = self.params.get(PHRASE_SEPARATOR, default_phrase_separator)
+        if phrase_uri is None:
+            phrase_uri = URIRef(self.base_uri + phrase_sep.join(phrase.split(" ")))
 
-def generate_windows(documents: dict = None, params: dict = {}):
+        q = (
+            """
+        SELECT distinct ?f (sum(?c) as ?num1)
+        WHERE
+        {
+            """
+            + phrase_uri.n3()
+            + """ rdf:value ?v . 
+            {
+                ?e ontolex:canonicalForm [ ontolex:writtenRep ?v ] .
+            }
+            UNION
+            {
+                ?e ontolex:otherForm [ ontolex:writtenRep ?v ] .
+            }
+            ?e ontolex:otherForm|ontolex:canonicalForm [ ontolex:writtenRep ?f ] .
+            ?p rdf:value ?f .
+            ?p nifvec:hasCount ?c .
+            ?p rdf:type nif:Phrase .
+        }
+        GROUP BY ?f
+        ORDER BY ?num1
+        """
+        )
+        results = [item for item in self.query(q)]
+        if len(results) > 0:
+            results = dict({r[0].value: r[1].value for r in results})
+        else:
+            results = dict()
+        return results
+
+    # setup a dictionary with phrases and their contexts to speed up
+    def load_vectors(
+        self,
+        documents: dict = None,
+        vectors: dict = None,
+        topn: int = 15,
+        includePhraseVectors: bool = True,
+        includeContextVectors: bool = False,
+        includeOtherForms: bool = False,
+    ):
+        """
+        Function to retrieve the vectors of phrases and context of a set of documents
+
+        """
+        if vectors is None:
+            vectors = dict()
+        params = {
+            WORDS_FILTER: {"data": {phrase: True for phrase in STOPWORDS}},
+            MIN_PHRASE_COUNT: 1,
+        }
+
+        documents = {key: preprocess(value) for key, value in documents.items()}
+
+        phrases = generate_document_phrases(documents=documents, params=params)
+        for phrase in phrases.keys():
+            if includePhraseVectors:
+                if vectors.get(phrase, None) is None:
+                    if includeOtherForms:
+                        vector = Counter()
+                        for form in self.find_otherForms(phrase):
+                            vector += self.phrase_contexts(form, topn=topn)
+                    else:
+                        vector = self.phrase_contexts(phrase, topn=topn)
+                    vectors[phrase] = vector
+            if includeContextVectors:
+                if vectors.get(context, None) is None:
+                    vectors[context] = self.context_phrases(context, topn=topn)
+        return vectors
+
+
+def document_vector(
+    documents: dict = None, vectors: dict = None, merge_dict: bool = False
+):
+    """
+    extract the phrases of a string and create dict of phrases with their contexts
+    """
+    params = {
+        WORDS_FILTER: {"data": {phrase: True for phrase in STOPWORDS}},
+        MIN_PHRASE_COUNT: 1,
+    }
+    documents = {key: preprocess(value) for key, value in documents.items()}
+    phrases = generate_document_phrases(documents=documents, params=params)
+    c = dict()
+    for phrase in phrases.keys():
+        if phrase not in vectors.keys():
+            logging.warning("Phrase " + phrase + " not found in vectors.")
+        c[phrase] = vectors.get(phrase, Counter())
+    #         c[context] = d.get(context, Counter())
+
+    if merge_dict:
+        c = merge_multiset(c)
+    return c
+
+
+def generate_document_contexts(
+    init_phrases: dict = None, documents: dict = None, params: dict = {}
+):
     """ """
-    windows = dict()
 
-    for key, document in documents.items():
-        for phrase, context, _ in generate_phrase_context(
-            document=document,
-            params=params,
+    logging.debug(".. generate document contexts started")
+
+    max_context_length = params.get(MAX_CONTEXT_LENGTH, default_max_context_length)
+    min_context_count = params.get(MIN_CONTEXT_COUNT, default_min_context_count)
+    min_phrasecontext_count = params.get(
+        MIN_PHRASECONTEXT_COUNT, default_min_phrasecontext_count
+    )
+    phrase_sep = params.get(PHRASE_SEPARATOR, default_phrase_separator)
+
+    init_contexts = defaultdict(lambda: defaultdict(lambda: defaultdict(set)))
+    for phrase, docs in init_phrases.items():
+        for doc, locs in docs.items():
+            for sent_idx, begin_idx, end_idx in locs:
+                sent = documents[doc][sent_idx]
+                if begin_idx - 1 >= 0 and end_idx + 1 <= len(sent):
+                    l = sent[begin_idx - 1]
+                    r = sent[end_idx]
+                    init_contexts[(l, r)][phrase][doc].add(
+                        (sent_idx, begin_idx, end_idx)
+                    )
+    del init_phrases
+
+    to_process_contexts = dict()
+    for d_context, d_phrases in init_contexts.items():
+        if len(d_phrases.keys()) > 1:
+            to_process_contexts[d_context] = (d_phrases, 1, 1)
+
+    # aggegrate results into contexts dict
+    final_contexts = defaultdict(Counter)
+    for d_context, d_phrases in init_contexts.items():
+        d_phrase_counter = Counter(
+            {
+                d_phrase: sum(len(loc) for loc in docs.values())
+                for d_phrase, docs in d_phrases.items()
+                if sum(len(loc) for loc in docs.values()) >= min_phrasecontext_count
+            }
+        )
+        if (
+            len(d_phrase_counter.keys()) > 0
+            and sum(v for v in d_phrase_counter.values()) >= min_context_count
         ):
-            if windows.get(phrase, None) is None:
-                windows[phrase] = Counter()
-            windows[phrase][context] += 1
+            final_contexts[d_context] = d_phrase_counter
+    del init_contexts
 
-    return windows
+    logging.debug(".... added initial contexts: " + str(len(final_contexts.keys())))
+    logging.debug(".... set of contexts to process: " + str(len(to_process_contexts)))
+
+    while to_process_contexts != dict():
+        new_contexts = defaultdict(lambda: defaultdict(lambda: defaultdict(set)))
+        for d_context, (
+            (d_phrases, left_size, right_size)
+        ) in to_process_contexts.items():
+            # print("evaluating "+str(d_context) + ": "+str(left_size)+", "+str(right_size))
+            for phrase, docs in d_phrases.items():
+                for doc, locs in docs.items():
+                    for sent_idx, begin_idx, end_idx in locs:
+                        sent = documents[doc][sent_idx]
+                        if d_context == (
+                            phrase_sep.join(sent[begin_idx - left_size : begin_idx]),
+                            phrase_sep.join(sent[end_idx : end_idx + right_size]),
+                        ):
+                            # right
+                            if (
+                                begin_idx - left_size >= 0
+                                and end_idx + right_size + 1 <= len(sent)
+                            ):
+                                l = phrase_sep.join(
+                                    sent[begin_idx - left_size : begin_idx]
+                                )
+                                r = phrase_sep.join(
+                                    sent[end_idx : end_idx + right_size + 1]
+                                )
+                                # print(".. (right) adding " +str((l, r)))
+                                new_contexts[(l, r)][phrase][doc].add(
+                                    (sent_idx, begin_idx, end_idx)
+                                )
+                            # left
+                            if (
+                                begin_idx - left_size - 1 >= 0
+                                and end_idx + right_size <= len(sent)
+                            ):
+                                l = phrase_sep.join(
+                                    sent[begin_idx - left_size - 1 : begin_idx]
+                                )
+                                r = phrase_sep.join(
+                                    sent[end_idx : end_idx + right_size]
+                                )
+                                # print(".. (left) adding " +str((l, r)))
+                                new_contexts[(l, r)][phrase][doc].add(
+                                    (sent_idx, begin_idx, end_idx)
+                                )
+
+        # determine contexts for further processing
+        to_process_contexts = dict()
+        for ((left_part, right_part)), d_phrases in new_contexts.items():
+            if (
+                len(d_phrases.keys()) > 1
+                and len(left_part.split(phrase_sep)) < max_context_length
+                and len(right_part.split(phrase_sep)) < max_context_length
+            ):
+                to_process_contexts[(left_part, right_part)] = (
+                    d_phrases,
+                    len(left_part.split(phrase_sep)),
+                    len(right_part.split(phrase_sep)),
+                )
+
+        # add new contexts to contexts
+        for d_context, d_phrases in new_contexts.items():
+            d_phrase_counter = Counter(
+                {
+                    d_phrase: sum(len(loc) for loc in docs.values())
+                    for d_phrase, docs in d_phrases.items()
+                    if sum(len(loc) for loc in docs.values()) >= min_phrasecontext_count
+                }
+            )
+            if (
+                len(d_phrase_counter.keys()) > 0
+                and sum(v for v in d_phrase_counter.values()) >= min_context_count
+            ):
+                final_contexts[d_context] = d_phrase_counter
+
+        logging.debug(".... added contexts: " + str(len(new_contexts.keys())))
+        logging.debug(
+            ".... set of contexts to process: " + str(len(to_process_contexts))
+        )
+
+    # create final phrases dict from contexts
+    phrases = Counter()
+    for d_context, d_phrases in final_contexts.items():
+        for phrase, value in d_phrases.items():
+            phrases[phrase] += value
+
+    logging.debug(".. generate document contexts finished")
+    logging.debug(".... total contexts: " + str(len(final_contexts.keys())))
+    logging.debug(".... total phrases: " + str(len(phrases.keys())))
+
+    return final_contexts, phrases
 
 
-def generate_phrase_context(
+def generate_document_phrases(documents: dict = None, params: dict = {}):
+    """
+    This function generates all phrases in the documents
+
+    :param documents: a dict with context.uri as keys and context.isString as values
+
+    :param params: a dict with parameters
+
+    """
+    logging.debug(".. generating document phrases")
+
+    min_phrase_count = params.get(MIN_PHRASE_COUNT, default_min_phrase_count)
+
+    # create a dict for each phrase that contain the phrase locations
+    phrases = defaultdict(lambda: defaultdict(set))
+    for context_uri, context_isString in documents.items():
+        for phrase, loc in generate_sentence_phrases(context_isString, params=params):
+            phrases[phrase][context_uri].add(loc)
+
+    # delete all phrases that occur less than then the min_phrase_count
+    to_delete = set()
+    for phrase, docs in phrases.items():
+        if sum(len(loc) for loc in docs.values()) < min_phrase_count:
+            to_delete.add(phrase)
+    for phrase in to_delete:
+        del phrases[phrase]
+
+    logging.debug(".... found phrases: " + str(len(phrases.keys())))
+
+    return phrases
+
+
+def generate_sentence_phrases(
+    sentences: list = None,
+    params: dict = {},
+):
+    """ """
+    phrase_sep = params.get(PHRASE_SEPARATOR, default_phrase_separator)
+    words_filter = params.get(WORDS_FILTER, None)
+    max_phrase_length = params.get(MAX_PHRASE_LENGTH, default_max_phrase_length)
+    for sent_idx, sentence in enumerate(sentences):
+        for word_idx, word in enumerate(sentence):
+            for phrase_length in range(1, max_phrase_length + 1):
+                if word_idx + phrase_length <= len(sentence):
+                    phrase_list = [
+                        sentence[word_idx + i] for i in range(0, phrase_length)
+                    ]
+                    phrase = phrase_sep.join(word for word in phrase_list)
+                    if words_filter is None:
+                        yield (
+                            phrase,
+                            (sent_idx, word_idx, word_idx + phrase_length),
+                        )
+                    else:
+                        # phrases may not start or end with one of the stopwords
+                        phrase_stop_words = [
+                            words_filter["data"].get(word.lower(), False)
+                            for word in [phrase_list[0], phrase_list[-1]]
+                        ]
+                        if not any(phrase_stop_words):
+                            yield (
+                                phrase,
+                                (sent_idx, word_idx, word_idx + phrase_length),
+                            )
+
+
+def preprocess(
     document: str = None,
     params: dict = {},
 ):
     """ """
-    phrase_sep = params.get(PHRASE_SEPARATOR, DEFAULT_PHRASE_SEPARATOR)
-    words_filter = params.get(WORDS_FILTER, None)
-    forced_sentence_split_characters = params.get(FORCED_SENTENCE_SPLIT_CHARACTERS, [])
-
-    max_phrase_length = params.get("max_phrase_length", 5)
-    min_left_length = params.get("min_left_length", 1)
-    max_left_length = params.get("max_left_length", 3)
-    min_right_length = params.get("min_right_length", 1)
-    max_right_length = params.get("max_right_length", 3)
-
-    tokenized_text = tokenize_text(document, forced_sentence_split_characters)
-    sentences = [[word["text"] for word in sentence] for sentence in tokenized_text]
-
-    # delete stopwords in sentence if enabled
-    for sentence in sentences:
-        sentence = ["SENTSTART"] + sentence + ["SENTEND"]
-        # perform regex selection of sentence
-        windows = {}
-        sentence = [word for word in sentence if re.match("^[0-9]*[a-zA-Z]*$", word)]
-        for idx, word in enumerate(sentence):
-            for phrase_length, left_length, right_length in product(
-                range(1, max_phrase_length + 1),
-                range(min_left_length, max_left_length + 1),
-                range(min_right_length, max_right_length + 1),
-            ):
-                if (
-                    idx
-                    >= max(
-                        1 if sentence[0] == "SENTSTART" else 0,
-                        left_length,
-                    )
-                    and idx
-                    <= len(sentence)
-                    - phrase_length
-                    - max(
-                        1 if sentence[-1] == "SENTEND" else 0,
-                        right_length,
-                    )
-                    and not (left_length == 0 and right_length == 0)
-                ):
-                    phrase_list = [sentence[idx + i] for i in range(0, phrase_length)]
-                    phrase = phrase_sep.join(word for word in phrase_list)
-                    left_context_list = [
-                        sentence[idx - left_length + i] for i in range(0, left_length)
-                    ]
-                    left_context = phrase_sep.join(word for word in left_context_list)
-                    right_context_list = [
-                        sentence[idx + phrase_length + i]
-                        for i in range(0, right_length)
-                    ]
-                    right_context = phrase_sep.join(word for word in right_context_list)
-                    context = (left_context, right_context)
-                    if context != ("SENTSTART", "SENTEND"):
-                        if words_filter is None:
-                            yield (
-                                phrase,
-                                context,
-                                (idx - left_length, idx + phrase_length + right_length),
-                            )
-                        else:
-                            phrase_stop_words = [
-                                words_filter["data"].get(word.lower(), False)
-                                for word in phrase_list
-                            ]
-                            left_phrase_stop_words = [
-                                words_filter["data"].get(word.lower(), False)
-                                for word in left_context_list
-                            ]
-                            right_phrase_stop_words = [
-                                words_filter["data"].get(word.lower(), False)
-                                for word in right_context_list
-                            ]
-                            if (
-                                not any(phrase_stop_words)  # and
-                                # (not all(left_phrase_stop_words)) and
-                                # (not all(right_phrase_stop_words))
-                            ):
-                                yield (
-                                    phrase,
-                                    context,
-                                    (
-                                        idx - left_length,
-                                        idx + phrase_length + right_length,
-                                    ),
-                                )
+    split_characters = params.get(FORCED_SENTENCE_SPLIT_CHARACTERS, [])
+    # tokenize documents into sentences
+    sentences = [
+        [word["text"] for word in sentence]
+        for sentence in tokenize_text(document, split_characters)
+    ]
+    # select alfanumeric and numeric tokens and add tokens for start and end of sentence
+    preprocessed = [
+        [
+            word
+            for word in ["SENTSTART"] + sentence + ["SENTEND"]
+            if re.match("^[0-9]*[a-zA-Z]*$", word)
+        ]
+        for sentence in sentences
+    ]
+    return preprocessed