+++ CHANGELOG.rst
@@ -61,3 +61,9 @@ Version 0.1.21 (2023-6-28)
 ==========================
 
 * Word filter added to NifVectorGraph
+
+Version 0.2.0 (2023-9-23)
+==========================
+
+* Reworked algorithm to create NifVectorGraph
+* Added functionality to create Lemon lexicon